Update prod.sh
Script does not ask for authorization before deleting files from directory.

Checks for directory, If the directory does not exist, it creates the directory before dumping all of the repo files in the directory.

If the directory exists, delete all files and subdirectories and replaces it with files from repo.